website_agent/requirements.txt
langchain
beautifulsoup4
requests
openai
python-dotenv
tqdm
playwright
streamlit